requirements.txt
streamlit>=1.37.0
ollama>=0.6.0
sentence-transformers>=5.0.0
feedparser>=6.0.0
newspaper3k>=0.2.8
lxml-html-clean>=0.4.0
pandas>=2.0.0
plotly>=5.0.0
requests>=2.32.0